Just do the rename
