+++ Cargo.toml
@@ -19,7 +19,7 @@ name = "deno_graph"
 all-features = true
 
 [features]
-type_tracing = ["deno_ast/transforms", "deno_ast/visit", "deno_ast/utils"]
+symbols = ["deno_ast/transforms", "deno_ast/visit", "deno_ast/utils"]
 
 [dependencies]
 anyhow = "1.0.43"
