+++ src/symbols/analyzer.rs
