+++ src/symbols/collections.rs
