+++ tests/specs/symbols/Basic.txt
