+++ tests/specs/symbols/Classes01.txt
