+++ tests/specs/symbols/DeclarationMerging01.txt
