+++ tests/specs/symbols/ExportAssignment01.txt
