+++ tests/specs/symbols/ExportDefault01.txt
