+++ tests/specs/symbols/ExportDefault02.txt
