+++ tests/specs/symbols/ExportDefault03.txt
