+++ tests/specs/symbols/ExportDefault04.txt
