+++ tests/specs/symbols/ExportDefaultLiteral01.txt
