+++ tests/specs/symbols/ExportDefaultLiteral02.txt
