+++ tests/specs/symbols/ExportDestructured.txt
