+++ tests/specs/symbols/ExportStar01.txt
