+++ tests/specs/symbols/ExportedDeclare01.txt
