+++ tests/specs/symbols/ExportedDeclare02.txt
