+++ tests/specs/symbols/Functions01.txt
