+++ tests/specs/symbols/ImportType01.txt
