+++ tests/specs/symbols/ImportType02.txt
