+++ tests/specs/symbols/Interfaces01.txt
