+++ tests/specs/symbols/Interfaces02.txt
