+++ tests/specs/symbols/ModuleExportNameStr01.txt
