+++ tests/specs/symbols/Namespaces01.txt
