+++ tests/specs/symbols/Namespaces02.txt
