+++ tests/specs/symbols/TsExternalModuleRef01.txt
