+++ tests/specs/symbols/TsExternalModuleRef02.txt
