+++ tests/specs/symbols/TsNamespaceExport01.txt
