+++ tests/specs/symbols/TsQualifiedName01.txt
