+++ tests/specs/symbols/TypeScriptTypesHeader.txt
