+++ tests/specs/symbols/TypesEntrypoint.txt
